Release

+++ music-cloud-web/src/main/web/src/app/components/home/home.component.ts
@@ -11,6 +11,7 @@ import { ChangeTrackDialog } from '../music/dialog/change-track-dialog';
 import { MatDialog } from '@angular/material';
 import { FileService } from '../../services/file.service';
 import { Observable } from 'rxjs';
+import {AddTrackToUserDialog} from '../music/dialog/add-track-to-user-dialog';
 
 @Component({
     selector: 'home',
@@ -27,6 +28,8 @@ export class HomeComponent implements OnInit {
     myRequest: any = [];
 
     myMusic: Track[] = [];
+    recommendedMusic: Track[] = [];
+    myPlaylists: any[] = [];
     response: TrackList;
     tracksLength : number = 10;
     pageEvent : any;
@@ -54,6 +57,7 @@ export class HomeComponent implements OnInit {
                         this.user = new User(data);
                         this.getPhoto();
                         this.loadTracksList(null);
+                        this.loadRecommendedTracksList();
                     },
                     error => {
                         if (error.status == 401) {
@@ -65,11 +69,36 @@ export class HomeComponent implements OnInit {
             this.user = new User(JSON.parse(this.shared.getStorage().getItem('loggedUser')));
             this.getPhoto();
             this.loadTracksList(null);
+            this.loadRecommendedTracksList();
         }
         } else {
             this.router.navigate(['login']);
         }
     }
+
+    addTrackToUser(id: number) {
+        this.trackService.addTrackToUser(this.user.id, id)
+            .subscribe(data => {
+                this.openTrackCreatedDialog(null);
+            });
+    }
+
+    addTrackToUserFromRec(id: number) {
+        this.trackService.addTrackToUser(this.user.id, id)
+            .subscribe(data => {
+                this.openTrackCreatedDialog(null);
+                this.loadTracksList(null);
+            });
+    }
+
+    openTrackCreatedDialog(response : any) : void {
+        const dialogRef = this.dialog.open(AddTrackToUserDialog, {
+            width: '250px',
+            data : response
+        });
+        dialogRef.afterClosed().subscribe(result => {
+        });
+    };
 
     isEmptyPhotoLink() {
         return this.user.email !== '' ? this.user.isEmptyPhotoLink() : false;
@@ -86,7 +115,7 @@ export class HomeComponent implements OnInit {
                     this.response = new TrackList(data);
                     this.tracksLength = this.response.allCount;
                     this.myMusic = this.response.tracks;
-                    this.loadAudioFiles();
+                    this.loadAudioFiles(this.myMusic);
                 });
         } else {
             this.trackService.getUserTracks(this.user.id, this.page, this.pageSize)
@@ -94,27 +123,42 @@ export class HomeComponent implements OnInit {
                 this.response = new TrackList(data);
                 this.tracksLength = this.response.allCount;
                 this.myMusic = this.response.tracks;
-                this.loadAudioFiles();
+                this.loadAudioFiles(this.myMusic);
             });
         }
     }
 
-    loadAudioFiles() {
-        for (let i = 0; i < this.myMusic.length; i++) {
-            this.loadFile(this.myMusic[i]);
+    loadRecommendedTracksList() {
+        this.trackService.getRecommendedUserTracks(this.user.id)
+            .subscribe(data => {
+                this.response = new TrackList(data);
+                this.recommendedMusic = this.response.tracks;
+                this.loadAudioFiles(this.recommendedMusic);
+            });
+    }
+
+    loadAudioFiles(tracks: Track[]) {
+        for (let i = 0; i < tracks.length; i++) {
+            this.loadFile(tracks[i]);
         }
     }
 
     deleteTrackFromUser(track : Track) {
         const dialogRef = this.dialog.open(DeleteTrackDialog, {
             width: '250px',
-            data : track
+            data : null
         });
         dialogRef.afterClosed().subscribe(result => {
-            this.trackService.deleteTrackFromUser(this.user.id, track.id)
-                .subscribe(() => {
-                    this.loadTracksList(null);
-                });
+            if (result) {
+                this.trackService.deleteTrackFromUser(this.user.id, track.id)
+                    .subscribe(() => {
+                        if (this.myMusic.length === 1) {
+                            this.deleteTrack(track.id, this.myMusic);
+                        } else {
+                            this.loadTracksList(null);
+                        }
+                    });
+            }
         });
     }
 
@@ -142,5 +186,13 @@ export class HomeComponent implements OnInit {
 
     loadFile(track: Track) {
         track.files = this.fileService.getUploadedTrack(track.filename);
+    }
+
+    deleteTrack(trackId: number, tracks: Track[]) {
+        let index = tracks.map(x => {
+            return x.id;
+        }).indexOf(trackId);
+
+        tracks.splice(index, 1);
     }
 }
